Update and rename company.py to employer.py
